Refactor CLIManager.h to update command descriptions with carriage return line endings for better formatting. Remove unused Helper.h file to streamline codebase. Modify MAE3Encoder.h to change return type of umToPulses from float to int32_t for improved precision. Enhance MotorSpeedController.h by adding a method to retrieve motor type. Update main.cpp to improve motor context management and command handling, including validation for target positions and motor indices. Improve debug output formatting across various files for consistency and readability.

+++ include/MAE3Encoder.h
@@ -112,11 +112,11 @@ public:
     }
 
     // Converts um to pulses
-    inline float umToPulses(float um)
+    inline int32_t umToPulses(float um)
     {
         float mm            = um / 1000.0f;
         float pulses_per_mm = FULL_SCALE / LEAD_SCREW_PITCH_MM;
-        return mm * pulses_per_mm;
+        return static_cast<int32_t>(mm * pulses_per_mm);
     }
 
     // Converts pulses to um
